fix: improve list files function
I went with using a temp file for the output of the list function so it
is piped through more. I also added a check for any files in devel that
are newer than their bin counterparts. They will be highlighted in red.
I also fixed some spelling errors. Finally, I renamed thins called "pad"
or "padding" to "col" and "separator."